+++ packages/common/src/milestones/models/create-milestone.model.ts
@@ -10,7 +10,6 @@ import {
 } from 'class-validator';
 import { CalendarInterval } from '@/calendar';
 import { CreateContentModel } from '@/content';
-import { PropertyType } from '@/models';
 
 @Exclude()
 export class CreateMilestoneModel extends CreateContentModel<CreateMilestoneModel> {