+++ packages/web/src/modules/habits/components/modals/EditHabitModal.vue
@@ -2,7 +2,10 @@
 import { CreateHabitModel, HabitModel, UpdateHabitModel } from '@lyvely/common';
 import { computed } from 'vue';
 import TagChooser from '@/modules/tags/components/TagChooser.vue';
-import { useContentEditModal } from '@/modules/content/composables/content-edit-modal.composable';
+import {
+  ContentEditModalEmits,
+  useContentEditModal,
+} from '@/modules/content/composables/content-edit-modal.composable';
 import { useHabitsService } from '@/modules/habits/services/habits.service';
 import { getCalendarPlanOptions } from '@/modules/calendar-plan';
 import { isTouchScreen } from '@/util';
@@ -17,7 +20,7 @@ export interface IProps {
 }
 
 const props = defineProps<IProps>();
-const emit = defineEmits(['update:modelValue']);
+const emit = defineEmits(ContentEditModalEmits);
 
 const { isCreate, showModal, model, validator, submit, status } = useContentEditModal<
   HabitModel,
@@ -35,7 +38,7 @@ const modalTitle = computed(() => {
 </script>
 
 <template>
-  <ly-modal v-model="showModal" :title="modalTitle" @submit="submit">
+  <ly-modal v-model="showModal" :title="modalTitle" @submit="submit" @cancel="$emit('cancel')">
     <template #preHeader><slot name="navigation"></slot></template>
     <ly-form-model
       v-model="model"