+++ packages/web/src/modules/milestones/components/modals/EditMilestoneModal.vue
@@ -1,7 +1,10 @@
 <script lang="ts" setup>
 import { computed } from 'vue';
 import TagChooser from '@/modules/tags/components/TagChooser.vue';
-import { useContentEditModal } from '@/modules/content/composables/content-edit-modal.composable';
+import {
+  ContentEditModalEmits,
+  useContentEditModal,
+} from '@/modules/content/composables/content-edit-modal.composable';
 import { CreateMilestoneModel, MilestoneModel, UpdateMilestoneModel } from '@lyvely/common';
 import { getCalendarPlanOptions } from '@/modules/calendar-plan';
 import { isTouchScreen } from '@/util';
@@ -16,7 +19,7 @@ export interface IProps {
 }
 
 const props = defineProps<IProps>();
-const emit = defineEmits(['update:modelValue']);
+const emit = defineEmits(ContentEditModalEmits);
 const store = useContentEditModal<MilestoneModel, CreateMilestoneModel, UpdateMilestoneModel>(
   props,
   emit,
@@ -33,7 +36,7 @@ const modalTitle = computed(() => {
 </script>
 
 <template>
-  <ly-modal v-model="showModal" :title="modalTitle" @submit="submit">
+  <ly-modal v-model="showModal" :title="modalTitle" @submit="submit" @cancel="$emit('cancel')">
     <template #preHeader><slot name="navigation"></slot></template>
     <ly-form-model
       v-model="model"